Hello there, this dracula theme fork is for people like me, that work on a server with minimal download permissions, where everything is made in house of downloaded by your boss approbal.
If you want dracula theme for zsh, you can just copy and paste the
content of this repo, bellow the paths and variables of your .zshrc
file.

I will keep this repo updated with the original one, so we have every
bit of vanilla configuration possible.

Hope this helps you.
Have a great day.